Fix incremental build with post-compile-weaving plugin
The post-compile-weaving plugin was incorrectly adding the compiler output directory to the _task inputs_ for modified compile tasks. This prevents the task from being
marked as UP-TO-DATE after an initial `clean compile`.

 
This fix removes the compiler output directory from the inputs of the modified
compile tasks. Instead, only files added to the `inpath` parameter are considered
to be task inputs. Both the compiler output and the configured `inpath` are still
passed to AJC via the `—inpath` property.

This PR fixes freefair/gradle-plugins#212.